refactor: deprecate the utility subpackage
